refactor(icons): inline glyph-source context into createIcon, drop useIconGlyphSource hook
