+++ packages/mobile/src/icons/index.ts
@@ -1,8 +1,11 @@
-export type { GlyphMap, IconGlyphResolverArgs, IconGlyphSource } from './createIcon';
-export { createIcon, DEFAULT_ICON_FONT_FAMILY } from './createIcon';
+export type {
+  GlyphMap,
+  IconGlyphResolverArgs,
+  IconGlyphSource,
+  IconGlyphSourceProviderProps,
+} from './createIcon';
+export { createIcon, DEFAULT_ICON_FONT_FAMILY, IconGlyphSourceProvider } from './createIcon';
 export * from './Icon';
-export type { IconGlyphSourceProviderProps } from './IconGlyphSourceContext';
-export { IconGlyphSourceProvider } from './IconGlyphSourceContext';
 export * from './LogoMark';
 export * from './LogoWordmark';
 export * from './SubBrandLogoMark';
